Move test specs next to source files
