+++ .mocharc.json
@@ -1,5 +1,5 @@
 {
     "extension": ["ts"],
-    "spec": ["test/**/*.ts"],
+    "spec": ["*/**/*.spec.ts"],
     "require": "ts-node/register"
 }
