+++ src/conversion/importExecutionResults/importExecutionResultsConverterServer.spec.ts
@@ -2,13 +2,13 @@
 
 import { expect } from "chai";
 import { readFileSync } from "fs";
-import { initOptions } from "../../../../src/context";
-import { ImportExecutionResultsConverterCloud } from "../../../../src/conversion/importExecutionResults/importExecutionResultsConverterCloud";
-import { InternalOptions } from "../../../../src/types/plugin";
-import { stubLogWarning } from "../../../constants";
-
-describe("the import execution results converter (cloud)", () => {
+import { stubLogWarning } from "../../../test/util";
+import { initOptions } from "../../context";
+import { InternalOptions } from "../../types/plugin";
+import { ImportExecutionResultsConverterServer } from "./importExecutionResultsConverterServer";
+describe("the import execution results converter (server)", () => {
     let options: InternalOptions;
+
     beforeEach(() => {
         options = initOptions({
             jira: {
@@ -28,7 +28,7 @@ describe("the import execution results converter (cloud)", () => {
         const result: CypressCommandLine.CypressRunResult = JSON.parse(
             readFileSync("./test/resources/runResult.json", "utf-8")
         );
-        const converter = new ImportExecutionResultsConverterCloud(options);
+        const converter = new ImportExecutionResultsConverterServer(options);
         const json = converter.convertExecutionResults(result);
         expect(json.tests[0].evidence).to.not.exist;
         expect(json.tests[1].evidence).to.be.an("array").with.length(1);
@@ -43,7 +43,7 @@ describe("the import execution results converter (cloud)", () => {
             readFileSync("./test/resources/runResult.json", "utf-8")
         );
         options.xray.uploadScreenshots = false;
-        const converter = new ImportExecutionResultsConverterCloud(options);
+        const converter = new ImportExecutionResultsConverterServer(options);
         const json = converter.convertExecutionResults(result);
         expect(json.tests).to.have.length(3);
         expect(json.tests[0].evidence).to.be.undefined;
@@ -56,7 +56,7 @@ describe("the import execution results converter (cloud)", () => {
             readFileSync("./test/resources/runResultProblematicScreenshot.json", "utf-8")
         );
         options.plugin.normalizeScreenshotNames = true;
-        const converter = new ImportExecutionResultsConverterCloud(options);
+        const converter = new ImportExecutionResultsConverterServer(options);
         const json = converter.convertExecutionResults(result);
         expect(json.tests[0].evidence[0].filename).to.eq("t_rtle_with_problem_tic_name.png");
     });
@@ -65,35 +65,35 @@ describe("the import execution results converter (cloud)", () => {
         const result: CypressCommandLine.CypressRunResult = JSON.parse(
             readFileSync("./test/resources/runResultProblematicScreenshot.json", "utf-8")
         );
-        const converter = new ImportExecutionResultsConverterCloud(options);
+        const converter = new ImportExecutionResultsConverterServer(options);
         const json = converter.convertExecutionResults(result);
         expect(json.tests[0].evidence[0].filename).to.eq("tûrtle with problemätic name.png");
     });
 
-    it("should use PASSED as default status name for passed tests", () => {
+    it("should use PASS as default status name for passed tests", () => {
         const result: CypressCommandLine.CypressRunResult = JSON.parse(
             readFileSync("./test/resources/runResult.json", "utf-8")
         );
-        const converter = new ImportExecutionResultsConverterCloud(options);
+        const converter = new ImportExecutionResultsConverterServer(options);
         const json = converter.convertExecutionResults(result);
-        expect(json.tests[0].status).to.eq("PASSED");
-        expect(json.tests[1].status).to.eq("PASSED");
+        expect(json.tests[0].status).to.eq("PASS");
+        expect(json.tests[1].status).to.eq("PASS");
     });
 
-    it("should use FAILED as default status name for failed tests", () => {
+    it("should use FAIL as default status name for failed tests", () => {
         const result: CypressCommandLine.CypressRunResult = JSON.parse(
             readFileSync("./test/resources/runResult.json", "utf-8")
         );
-        const converter = new ImportExecutionResultsConverterCloud(options);
+        const converter = new ImportExecutionResultsConverterServer(options);
         const json = converter.convertExecutionResults(result);
-        expect(json.tests[2].status).to.eq("FAILED");
+        expect(json.tests[2].status).to.eq("FAIL");
     });
 
     it("should use TODO as default status name for pending tests", () => {
         const result: CypressCommandLine.CypressRunResult = JSON.parse(
             readFileSync("./test/resources/runResultPending.json", "utf-8")
         );
-        const converter = new ImportExecutionResultsConverterCloud(options);
+        const converter = new ImportExecutionResultsConverterServer(options);
         const json = converter.convertExecutionResults(result);
         expect(json.tests[0].status).to.eq("TODO");
         expect(json.tests[1].status).to.eq("TODO");
@@ -101,14 +101,14 @@ describe("the import execution results converter (cloud)", () => {
         expect(json.tests[3].status).to.eq("TODO");
     });
 
-    it("should use FAILED as default status name for skipped tests", () => {
+    it("should use FAIL as default status name for skipped tests", () => {
         const result: CypressCommandLine.CypressRunResult = JSON.parse(
             readFileSync("./test/resources/runResultSkipped.json", "utf-8")
         );
-        const converter = new ImportExecutionResultsConverterCloud(options);
+        const converter = new ImportExecutionResultsConverterServer(options);
         const json = converter.convertExecutionResults(result);
-        expect(json.tests[0].status).to.eq("FAILED");
-        expect(json.tests[1].status).to.eq("FAILED");
+        expect(json.tests[0].status).to.eq("FAIL");
+        expect(json.tests[1].status).to.eq("FAIL");
     });
 
     it("should be able to use custom passed statuses", () => {
@@ -116,7 +116,7 @@ describe("the import execution results converter (cloud)", () => {
             readFileSync("./test/resources/runResult.json", "utf-8")
         );
         options.xray.statusPassed = "it worked";
-        const converter = new ImportExecutionResultsConverterCloud(options);
+        const converter = new ImportExecutionResultsConverterServer(options);
         const json = converter.convertExecutionResults(result);
         expect(json.tests[0].status).to.eq("it worked");
         expect(json.tests[1].status).to.eq("it worked");
@@ -127,7 +127,7 @@ describe("the import execution results converter (cloud)", () => {
             readFileSync("./test/resources/runResult.json", "utf-8")
         );
         options.xray.statusFailed = "it did not work";
-        const converter = new ImportExecutionResultsConverterCloud(options);
+        const converter = new ImportExecutionResultsConverterServer(options);
         const json = converter.convertExecutionResults(result);
         expect(json.tests[2].status).to.eq("it did not work");
     });
@@ -137,7 +137,7 @@ describe("the import execution results converter (cloud)", () => {
             readFileSync("./test/resources/runResultPending.json", "utf-8")
         );
         options.xray.statusPending = "still pending";
-        const converter = new ImportExecutionResultsConverterCloud(options);
+        const converter = new ImportExecutionResultsConverterServer(options);
         const json = converter.convertExecutionResults(result);
         expect(json.tests[0].status).to.eq("still pending");
         expect(json.tests[1].status).to.eq("still pending");
@@ -150,9 +150,9 @@ describe("the import execution results converter (cloud)", () => {
             readFileSync("./test/resources/runResultSkipped.json", "utf-8")
         );
         options.xray.statusSkipped = "omit";
-        const converter = new ImportExecutionResultsConverterCloud(options);
+        const converter = new ImportExecutionResultsConverterServer(options);
         const json = converter.convertExecutionResults(result);
-        expect(json.tests[0].status).to.eq("FAILED");
+        expect(json.tests[0].status).to.eq("FAIL");
         expect(json.tests[1].status).to.eq("omit");
     });
 
@@ -161,7 +161,7 @@ describe("the import execution results converter (cloud)", () => {
             readFileSync("./test/resources/runResultUnknownStatus.json", "utf-8")
         );
         const stubbedWarning = stubLogWarning();
-        const converter = new ImportExecutionResultsConverterCloud(options);
+        const converter = new ImportExecutionResultsConverterServer(options);
         const json = converter.convertExecutionResults(result);
         expect(stubbedWarning).to.have.been.calledWith(
             "Unknown Cypress test status: 'broken'. Skipping result upload for test \"TodoMVC hides footer initially\"."
@@ -176,7 +176,7 @@ describe("the import execution results converter (cloud)", () => {
         const result: CypressCommandLine.CypressRunResult = JSON.parse(
             readFileSync("./test/resources/runResult.json", "utf-8")
         );
-        const converter = new ImportExecutionResultsConverterCloud(options);
+        const converter = new ImportExecutionResultsConverterServer(options);
         const json = converter.convertExecutionResults(result);
         expect(json.tests).to.have.length(3);
         expect(json.tests[0].testInfo.steps).to.have.length(1);
@@ -192,7 +192,7 @@ describe("the import execution results converter (cloud)", () => {
             readFileSync("./test/resources/runResult.json", "utf-8")
         );
         options.xray.steps.update = false;
-        const converter = new ImportExecutionResultsConverterCloud(options);
+        const converter = new ImportExecutionResultsConverterServer(options);
         const json = converter.convertExecutionResults(result);
         expect(json.tests).to.have.length(3);
         expect(json.tests[0].testInfo.steps).to.be.undefined;
@@ -204,7 +204,7 @@ describe("the import execution results converter (cloud)", () => {
         const result: CypressCommandLine.CypressRunResult = JSON.parse(
             readFileSync("./test/resources/runResultLongBodies.json", "utf-8")
         );
-        const converter = new ImportExecutionResultsConverterCloud(options);
+        const converter = new ImportExecutionResultsConverterServer(options);
         const json = converter.convertExecutionResults(result);
         expect(json.tests[0].testInfo.steps[0].action).to.eq(`${"x".repeat(7997)}...`);
         expect(json.tests[1].testInfo.steps[0].action).to.eq(`${"x".repeat(8000)}`);
@@ -216,7 +216,7 @@ describe("the import execution results converter (cloud)", () => {
             readFileSync("./test/resources/runResultLongBodies.json", "utf-8")
         );
         options.xray.steps.maxLengthAction = 5;
-        const converter = new ImportExecutionResultsConverterCloud(options);
+        const converter = new ImportExecutionResultsConverterServer(options);
         const json = converter.convertExecutionResults(result);
         expect(json.tests[0].testInfo.steps[0].action).to.eq("xx...");
         expect(json.tests[1].testInfo.steps[0].action).to.eq("xx...");
