+++ src/util/base64.spec.ts
@@ -1,7 +1,7 @@
 import { expect } from "chai";
-import { encodeFile } from "../../../src/util/base64";
+import { encodeFile } from "./base64";
 
-describe("the encoding utility", () => {
+describe("the base64 utility", () => {
     it("should be able to encode png files to base64", () => {
         const encodedString = encodeFile("./test/resources/turtle.png");
         expect(encodedString).to.have.length.greaterThan(0);
