+++ src/util/zip.spec.ts
@@ -2,8 +2,8 @@ import chai, { expect } from "chai";
 import chaiAsPromised from "chai-as-promised";
 import fs from "fs";
 import path from "path";
-import { unzip } from "../../../src/util/zip";
-import { getTestDir } from "../../constants";
+import { getTestDir } from "../../test/util";
+import { unzip } from "./zip";
 
 // Enable promise assertions.
 chai.use(chaiAsPromised);
